chore: clear the compiler warnings and let CI fail on new ones
A build that always prints warnings is a build where a new one is
invisible, which is why CI could not use -Werror: main would have
failed on its own. -Wall -Wextra now come out clean in both build
types, they are on for everyone, and CI configures with SST_WERROR=ON
so the next one stops a pull request instead of scrolling past.

Most of it says out loud what the code already meant. Five switch cases
that fall into the next on purpose say so with a marker the compiler
can read, two of them replacing comments that said it only to human
readers. A count of commands is signed, because everything that walks
it is. A variable that was set and never read is gone.

Four were real, though, and they were all the same shape: a value read
before anything had written it.

Loading a saved game never checked a single one of its thirteen reads.
A file that stopped short -- an interrupted freeze, a truncated copy --
was read into the game state and played anyway; in the worst shape,
with a password that had never been read at all. It now says so and
asks again, the way a file that was not there already did. Refusing it
means putting back everything the partial read had touched, not just
the password: the length and skill live in the same struct, and leaving
them behind handed the player the damaged file's game instead of the
one they went on to ask for, and swallowed the two answers they were
about to type.

A move too short to take a single step skipped the loop that walks the
ship across the quadrant -- and the enemy-distance sums after it read
the position that loop would have written. They now start from where
the ship is, which for a move that never happened is where it stays.
This is the one change ordinary play can see: the sums were being taken
from sector 0-0.

Capturing a Klingon picked its target with a search that returned
nothing at all if every enemy present was a Romulan. Unreachable behind
the caller's own check, but it indexed arrays with the result.

A base attack by the Super-commander saved where a commander's attack
had been pointed, in a local, and put it back in a later call that had
never set it. It lives across calls now. The attack that fizzles
because the base is already gone says so with a zero rather than
leaving the last attack's coordinates to be restored over a base that
no longer exists -- which would have frozen the Super-commander for the
rest of the game.

Ordinary play is byte-identical to main across forty seeded journeys in
both build types, freeze and thaw included. The short-move path is not,
and that is the fix.

The seeded battery gains a case for the damaged save file, and a
short-move jump so that path is played at all. The battery asserts the
setup questions come back, which stands in for the state having been
cleared rather than checking it directly.

Fixes #14

Entire-Checkpoint: e46dbe373750

+++ sst.c
@@ -100,7 +100,8 @@ static char *commands[] = {
  
 };
 
-#define NUMCOMMANDS (sizeof(commands)/sizeof(char *))
+/* Signed, because every loop below counts through it with an int. */
+#define NUMCOMMANDS ((int)(sizeof(commands)/sizeof(char *)))
 
 static void listCommands(int x) {
 	prout("   SRSCAN    MOVE      PHASERS   CALL\n"
